+++ include/porto/raytracer.h
@@ -13,7 +13,7 @@ namespace porto
     public:
         Camera cam;
         Scene scene;
-        Raytracer(int width = 800, int height = 600, float vfov = 90,
+        Raytracer(int width = 800, int height = 600, double vfov = 90,
                 Vec3 origin = Vec3(0,0,0));
         ~Raytracer();
         Vec3 getPixel(int x, int y);
