+++ src/camera.cpp
@@ -2,11 +2,11 @@
 #include <porto/vec3.h>
 #include <math.h>
 
-void porto::Camera::setup(float vfov, float aspect)
+void porto::Camera::setup(double vfov, double aspect)
 {
-    float theta = vfov*M_PI/180;
-    float half_height = tan(theta/2);
-    float half_width = aspect * half_height;
+    double theta = vfov*M_PI/180;
+    double half_height = tan(theta/2);
+    double half_width = aspect * half_height;
     lower_left_corner = Vec3(-half_width, -half_height, -1.0);
     horizontal = Vec3(2*half_width, 0.0, 0.0);
     vertical = Vec3(0.0, 2*half_height, 0.0);
@@ -15,38 +15,38 @@ void porto::Camera::setup(float vfov, float aspect)
     this->aspect = aspect;
 }
 
-porto::Camera::Camera(float vfov, float aspect)
+porto::Camera::Camera(double vfov, double aspect)
 {
     setup(vfov, aspect);
 }
 
-porto::Camera::Camera(float vfov, float width, float height)
+porto::Camera::Camera(double vfov, double width, double height)
 {
-    float aspect = width / height;
+    double aspect = width / height;
     setup(vfov, aspect);
 }
 
-porto::Ray porto::Camera::getRay(float u, float v) {
+porto::Ray porto::Camera::getRay(double u, double v) {
     return porto::Ray(origin, lower_left_corner + u*horizontal + v*vertical - origin);
     // to add diafragma add to origin random number
 }
 
-void porto::Camera::setVfov(float vfov)
+void porto::Camera::setVfov(double vfov)
 {
     setup(vfov, aspect);
 }
 
-float porto::Camera::getVfov()
+double porto::Camera::getVfov()
 {
     return vfov;
 }
 
-void porto::Camera::setAspect(float aspect)
+void porto::Camera::setAspect(double aspect)
 {
     setup(vfov, aspect);
 }
 
-float porto::Camera::getAspect()
+double porto::Camera::getAspect()
 {
     return aspect;
 }
